feat: separate footer and header into own components
