+++ app/footer.tsx
@@ -1,0 +1,6 @@
+export function Footer() {
+    return (
+        <footer className="border-stone-900">
+        </footer>
+    )
+}
